core. Store endpoint in IP format
If a VPN endpoint is given in host name form, resolve it to IP address with
DNS.

The DNS service might be not reachable during system boot. Configuring
the WireGuard device with a plain IP address ensures the interface is
started despite DNS failures.

If the node IP address changes over the time it is better to not set the
endpoint at all.