core. Run custom update scripts of new image
The update-module/50run_scriptdir action step, provided by the core as default, can run
the executables under a well-known directory:

    ${AGENT_INSTALL_DIR}/update-module.d

This solution ensures that any executable installed with the new module
image actually has a chance to run.

Scripts execution occurs in alphabetical order. Execution breaks when a
script exits non-zero. In that case the whole action is aborted.