Fix cluster/favorites clean up
Store the image_name in a dedicated Redis key. The key is created when a
module instance is successfully added (instead of creating it early).

The module/{module_id} HASH is removed, as it correspond to a key
namespace prefix and might lead to confusion.

Clean up cluster/favorites in remove-module.