core. module-backup generic command implementation
Implement backup in a single command (a Podman+Restic wrapper). The
command is designed to run in the agent environment and can be activated
from a Systemd timer unit.

Podman volumes are mounted under /srv/volumes of the Restic container
Restic container working directory is /srv, all patterns are relative to
it.

Config files (optional):

    etc/state-include.conf
    etc/state-exclude.conf

Files always included in the backup:

    state/environment

Common wrapper logic is implemented in the agent.run_restic() function.