core. Use rclone binary from docker.io
The rclone-wrapper agent command is available to ease rclone invocation.

The wrapper expects the first argument is a repository ID, or a backup
ID. It retrieves from the local Redis DB the credentials and the remote
path to access the remote. The wrapper also replaces any occurrence of
"REMOTE_PATH" with the good value, obtained from the repository "url"
field.

If DEBUG=1 is defined in the environment, the full Podman+Rclone
invocation is dumped to stderr.