samba. Create a dummy ethernet device on the fly
The samba service wants a static IP address to start correctly. The
`samba-network` script ensures that the IPADDRESS is always available on
the system, creating a dummy interface when needed, every time the samba
service starts.

As restore-module action precondition, run samba-network up to ensure
IPADDRESS exists on the system.