traefik: use `.test` tld for selfsigned cert
If the node's FQDN is used in the default selfsigned certificate, when
a valid certificate for the node's FQDN is requested using the
`set-certificate` action, `traefik` will not request a new certificate
via Let's Encrypt, but will use the already present selfsigned
certificate.
This is caused by the fact that in the system is already present a
certificate for the requested domain (in the selfsigned certificate) and
to avoid this problem the tld `.test` is used in the selfsigned
certificate.
The `CN` of the selfsigned certificate will be `FQDN + .test`