Persist authorization labels to refresh them
The org.nethserver.authorizations label of each module is stored in
Redis to re-evaluate it each time a new module is added to the cluster.
Key prefix is cluster/authorizations/

Also store the module-node association in a safe key:
cluster/module_node.

Added keys and also roles/ now are cleaned up in remove-module.