Allow addition of pip packages (#322)
This Python venv flag allows to install packages with the `--user` flag.
E.g. `pip install --user Crypto`

It allows rootless modules to add more Python packages to the base 
`pyenv` agent environment: they are installed under the module home
 directory.

The recommended place to run such command is the `create-module` 
action, in an early step (e.g.  05pip_install). Uninstall is not needed because
files are removed with the whole module home directory.