Override replace flag for restore-module
If a module backup is restored in a cluster with no instances of that
module, or if there is not an instance with the same MODULE_UUID, the
restore-module action is invoked with "replace=true", to preserve the
original MODULE_UUID from backup.

The UI actually disables the Replace checkbox under the same conditions,
preventing MODULE_UUID to be preserved.