Two steps installation process
Separate the installation of an NS8 system into two steps, in the first
step will be installed all the software needed and in the second step
will be configured the system and will be created the passwords, keys,
and tokens.

All the software is installed by the `install-core.sh` and the
`install-finalize.sh` script will configure the systems and generate the
credentials.

This will ease the creation of a generic image that can be used as a
base for future parametrized installations.

The feature controlled is by the environment variable
`NS8_TWO_STEPS_INSTALL`