Avoid ldap3 library error against NS7 OpenLDAP
The ldap3 Python library uses the server provided schema to check search
query attributes. If the attribute name is unknown in the schema it raises a
LDAPAttributeError exception.

As disabing the schema would impact the library return value format,
check if the query attributes are in the schema and filter out unknown
attributes to avoid the error.