actions: build NS8 images
This workflow will create disk images that can be used to quickly create
a new instance of an NS8 node. The images a based on all the supported
distros:

* Debian 11
* Centos Stream 9
* Rocky Linux 9
* Alma Linux 9

The images use the QCOW2 format and are compacted and compressed.

The default root password is Nethesis,1234, and will be asked to change
it on the first login.