OTP login protocol refactor
The /api/login POST route checks if 2FA is needed or not, according to
Redis HGET user/<username> 2fa. If 2FA is needed, but OTP is not given a
401 is returned.  The failure reason is in the response message.

If login is successful, store in the "2fa" JWT token claim a boolean
value:

- true if JWT was obtained from /api/login with a valid OTP
- false if JWT was obtained with username+password credentials

In any case a JWT token grants access to /api/* until it expires.

In the UI login form logic, 401 response messages are properly handled
to decide if a OTP prompt must be shown and OTP must be sent to
/api/login.

As the Python agent.tasks library never uses 2FA, no changes are needed
there.