Fix ACL disk sync
Every time Redis ACLs are changed they must be saved in the cluster node
disk (ACL SAVE) and propagated to worker nodes.

This commit adds ACL SAVE in the cluster.grants.save_acls() function, to
ensure we do not forget to do it in cluster API implementation. As
consequence, sparse ACL SAVE commands are removed.