install. Fix Rocky Linux installation
The RPM key is not imported on a pristine Rocky Linux installation,
until the first RPM is installed.

If we disable official Rocky repos without importing the key DNF does
not how to find it.

This commit imports the key in the RPM DB forcibly, so it is always
found.