chore: Add check for /usr/local/bin in PATH environment variable
The code changes include adding a check for the presence of /usr/local/bin in the PATH environment variable. If /usr/local/bin is not found in the PATH, an error message is displayed and the script exits. This change ensures that the necessary directory is included in the PATH for proper execution of the script.