Run the script only at boot
The unit is pulled in by the Redis unit, and runs before it at boot.
However it must run before containers are started, only one time at
boot, otherwise the removal of merged/ dirs breaks the overlay
filesystem.

Prevent the script to be invoked multiple times, if Redis is restarted.