update-module. Mandatory module_url parameter
The image_url must be a good update for all instances passed as
parameter. As it is difficult to check if it is appropriate or not, or
to calculate an image_url that fits all instances (if not given), we
drop the automatic latest version lookup and we expect a good one is
passed.