Avoid deadlock after SIGUSR1 is received
Listen for new tasks until workers finish. A subtask created by a parent worker
needs to be processed to avoid entering a deadlock state.

When SIGUSR1 is received, tasks are still pulled from the Redis queue until all
running workers finish. At that point, the Redis BRPOP goroutine
is canceled, and we can exit the actions loop.

This commit also protects the shared taskCancelFunctions map with a Mutex
because it is accessed by many goroutines simultaneously.