fix(create-cluster): use Traefik config of workers
Traefik API connections originate only from the localhost, access from
cluster VPN is not needed.

The cluster VPN is not automatically inserted in Traefik configuration
of worker nodes. As result there is a configuration difference between the original
cluster leader and other nodes. This commit wants to remove this
conflicting configuration.