feat: new user-domain-changed event format
Add the `domains` attribute to event's payload. The value is a list of
domains. The old `domain` attribute is still present and will equal to
a random list element.

This new event format with multiple domains is necessary to reduce the
size of the event burst generated by Ldapproxy when a node joins the
cluster.